feat: Same syntax highlight theme between prism and codemirror
When user edit a query in cross tracker search widget, the syntax highlighting should be the same than the one used to display the query in readonly mode, so that the user is not confused.

Bring back the `dedent` when user use a closing parenthesis.

Close request #39683: Syntax highlight query should be the same between write and read modes

Change-Id: I8825d9014cec8532b851c816c6cb65edb0db6e7b