feat: add of existing section is always available
When the configured tracker for an artidoc does not contain a
submittable section (e.g no submit rights for title or description) then
current users are not able to add a new section.

However there is no reason to prevent them to add a section from an
existing artifact.

Part of story #39369: Add section from existing artifact

Change-Id: I530ce2999db39d9a10a04bf3e5921e1a1c513168