feat: Introduce @tuleap/prism-language-tql
Closes request #39671 TQL syntax highlighting with Prism

Introduce js library @tuleap/prism-language-tql to have syntax
highlighting for TQL.

Comments are not valid TQL but there are available here as it can be
usefull in a codeblock.

*Testing:*

In a text field of an artifact, add a tql codeblock with some TQL inside.
After saving, the code should be colored.

Change-Id: Ic5db734d94f6832fa5c571756f3c64d5cc616b42