feat: retrieve PlanConfiguration from Template project
part of request #39352 Program add to top backlog post-action should
ignore Teams

How to test:
- Set up a Program project (A) that will serve as template. It should
have no Team project linked to it.
- Create a new B project from a project that you are admin of, and
select A as template. Create the project.
- In program_management_syslog, you should see a log line detailing the
Plan configuration retrieved from database for the A template project.
In later contributions, it will be modified by the mapping from the
event, turned into a NewPlan for project B and saved to DB.

Why?
Instead of dealing only with arrays and primitives, we let the DAO
return a full Domain object. In later contributions, we can modify its
data (with the mapping) and create a new NewPlan for the new project to
save.

Change-Id: Ie5c11867c89c1653ab02c4384a0491ce87069c53