refactor: rename Plan -> NewPlan
part of request #39352 Program add to top backlog post-action should
ignore Teams

No functional change.

Why?
The previous "Plan" object contains contraints that apply only on new
plan configurations, not existing configurations. For example, it has no
Program identifier attached to it (because it might not exist yet). We
apply the convention of prefixing by "New" to make it clearer that it's
a new creation and as such, cannot represent an existing state.

Change-Id: Ie4e8473089906245c9891c8788f0c57bc16bfb9b