feat: Syntax highlight query in reading mode
When Cross Tracker Search widget is in display mode, we should have
syntax highlighting for the query so that it is more easy for the user
to parse the query.

Took the opportunity to better distinquish `@title` variable in the
query `SELECT @title FROM`.

Part of request #39680: Syntax highlight query in reading mode

Change-Id: Ibaa5e6b8aa8ed709f0ef6a8b2545bb9ba6ddac15