fix: Highlighting should be kept on cancel
Go to a cross tracker search widget with a saved expert query.
The query is syntax highlighted in readonly mode.
Edit the query and run the query.
The query is still syntax highlighted in readonly mode.
Click on [Cancel] to retrieve the saved query.
The query SHOULD be syntax highlighted.

The issue was due to the custom element that is not aware of content
change.

Part of request #39680: Syntax highlight query in reading mode

Change-Id: Iec20997faa4f06b83bdfd6e3c6a2c6a2597e50c3