refactor: unit test insertion of link popovers
story #38627 : Choice of a new Rich Text Editor

Note:
22 unit-tests have been added to cover as much as possible the insertion of link popovers.

In order to ease the unit tests, I had to split the code in several
class-like object, with their corresponding types. It allows us to stub
the prose-mirror stuff in the unit tests of our business code, which is
kind of cool given the difficulty to manipulate EditorState and
EditorView oustside a real editor.

-- no functional change expected.

Change-Id: Ibacf8df0b959b1fc71b309d276724e248d7e7681