feat: add subscript support in prose-mirror editor
Add subscript mark support in prose-mirror editor.
A subscript button in the toolbar allows the user to apply subscript style on selected text.
There is also a shortcut available Ctrl + ,
A plugin has been created to prevent the user to be blocked in a mark while they writing. (Note: for now, it's only applied on subscript mark, to support other marks, you can add it in "MARK_TO_CLOSE" array)
If the user writes at the end of the line in a mark, they can press "space" to close the current mark and come back to a normal style, like this :

Issue :
I can't <sub>123 write the end of my sentence after a mark !</sub>

and have this instead :

I can <sub>123</sub> write the end of my sentence after a mark !

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I60d21f9670fd0a71a074123e7e03c89348926a08