feat: Show the same empty state for default reports
part of story #39685 Order artifact according to a field

How to test:
- Create a brand new cross-tracker widget. It will be in default mode.
The empty state has an illustration and mentions that the query is empty
and that the user should click on the "No selected element" label.
- When you write a query without selecting trackers, the empty state
will mention to select trackers.
- When you write a query that matches no artifact (for example @id < 1),
the empty state appears.
- No functional change when report is in expert mode. The empty state
keeps appearing when no artifact matches the query.

Change-Id: I6992dfe1c6ea865e85d0e75b2762e0a8942840ed