refactor: rename Plan -> PlanConfiguration
part of request #39352 Program add to top backlog post-action should
ignore Teams

No functional change.

Why?
"A plan" can be a map or a strategy, or a verb "to plan". Since we deal
with the data necessary in order "to plan" features in program
increments, we rename it to "plan configuration". The data is not "a
plan".

Change-Id: I97cd8f949962b1e4f54bbf3da1dd8d4307ed8b1b