fix: quote icon flickering
By default, prose-mirror hide quote icon when the selected node is not compatible with quoteblock.
So I built a custom icon, that is always active.
I also rewrite the default behavior of quote because it was not compatible with list block.

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I0f9d5ac9168d204af4a04a45b37c7859aee21795