refactor: rethink popover buttons inclusion
part of story #38627 : Choice of a new Rich Text Editor

Currently, the design of the popover buttons inclusion is not really
good:
- Each button element must extend a "type" property, but needs to
omit it in their component definition (in order to not have to define a
readonly property).
- We need a button-creator to render each button. It has to somehow pass
  the props to the buttons so they can render themselves.

In short, this is kind of a mess.

We had to establish this strategy because there were a bug in hybrids
preventing us to pass children components their properties ([see GH
issue](https://github.com/hybridsjs/hybrids/issues/273)).

Now it has been corrected, and that hybrids has been successfully upgraded to version 9.1.6 in the whole codebase, then we can refactor the current popover buttons inclusion system to make it simpler
and more convenient to use.

-- no functional change expected

Change-Id: Iff689add5532b1d0ea6a5ae15f2b7218402d5e4f