feat: Apply tracker mapping to Program Increment tracker
part of request #39352 Program add to top backlog post-action should
ignore Teams

How to test:
- Set up a Program project (A) that will serve as template.
- Create a new B project from a project that you are admin of, and
select A as template. Create the project.
- In program_management_syslog, you should see a log line detailing the
new program ID (project B), the mapped tracker ID for Program Increment
tracker for project B, and the label and sub-label for Program Increment
retrieved from Project A.

Why?
This is a first step towards applying the tracker mapping for all
configured tracker for Program configuration. In next steps, we will
apply the mapping for iterations and trackers that can be planned
(Features and Enablers).

Change-Id: Iefe8d997e56bcefc1e56479a57b6d09d50d7e41d