feat: Apply tracker mapping to Iteration tracker
part of request #39352 Program add to top backlog post-action should
ignore Teams

How to test:
- Set up a Program project (A) that will serve as template.
- Configure a tracker and custom labels for Iterations in Program A.
- Create a new B project from a project that you are admin of, and
select A as template. Create the project.
- In program_management_syslog, you should see a log line detailing the
new program ID (project B), the mapped data for Program Increment, and
the mapped tracker ID for Iteration tracker in project B, and the label
and sub-label for Iterations retrieved from Project A.
- When there is no Iteration tracker set up in project A, only empty
strings are logged.

Why?
This is a next step towards applying the tracker mapping for all
configured trackers for Program configuration. In next steps, we will
apply the mapping for trackers that can be planned (Features and
Enablers).

Change-Id: I5173b772ab73cb304a6b1065db4ac0c8d77767b7