feat: add superscript support in prose-mirror editor
Add superscript mark support in prose-mirror editor.
A superscript button in the toolbar allows the user to apply superscript style on selected text.
There is also a shortcut available Ctrl + .
If the user writes at the end of the line in a mark, they can press "space" to close the current mark and come back to a normal style.

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I519fe501f54e6143c0c3413c444d0b263359ed38