feat: Apply tracker mapping to Trackers that can be planned
part of request #39352 Program add to top backlog post-action should
ignore Teams

How to test:
- Set up a Program project (A) that will serve as template.
- Configure trackers that can be planned in Program Increments (for
example Features and Enablers).
- Create a new B project from a project that you are admin of, and
select A as template. Create the project.
- In program_management_syslog, you should see a log line detailing the
new program ID (project B), the mapped data for Program Increment and
Iteration, and the mapped tracker IDs for each tracker that can be
planned.

Why?
This is a next step towards applying the tracker mapping for all
configured trackers for Program configuration. In next steps, we will
apply the mapping for user groups that can prioritize.

Change-Id: I481a0b1dfbaa702bcebb92d2d4d5d4949b505717