design: apply box-shadow on toolbar when stuck
Toolbar in artidoc is marked as sticky.

The box-shadow (sticky-header-shadow) is here to indicate to the user
that some content is being scrolled under it. Therefore this box-shadow
should be displayed only when the toolbar is "stuck".

Implementation note: the combination of plateform banner, project
banner, and sticky header make it difficult to use IntersectionObserver
so we use scroll event instead.

Additionally, the position of the toolbar is now adapted accordingly to
the presence of project banner, platform banner, or both.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: I8f14083dbaaac11da46b66145649b5ff38b2fe76