fix: Greater whitespace in PR homepage
The creation date is too close to the "Created by" label. Having a
greater whitespace brings breathability and is more comfortable. (See
request #39719 if you need a screenshot of the issue).

Took the opportunity to remove `div` inside `span` in this file for the
sake of html semantic. `span` element accepts only phrasing content,
`div` is a flow content and therefore should not be under a `span`.

Closes request #39719: Lack of whitespace between date and Created by in PR homepage

Change-Id: I9c9c070383945344ddd70b997c4248e1790d03f8