fix: when selection has more than a link, unlink should remove all links
When the user selects multiple links and clicks on remove links button, it should remove all selected links.
When the user selects text with links, we search the index of the first link and the end of the last link to update the selection positions to be sure to include all links.
The remove-links has been moved from src/helpers to link/helpers.

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I8d52163848bed0032897d49b7ad2a2a40407b007