fix: tables width in ttm docx generation
Go to a TTM campaign, and download it as a document.

The generated document is well formatted in LibreOffice 💪

However once imported in Google Docs or in MS Office, the matrix or the
artifact links tables are screwed: some columns are very very small and
the content is written vertically 👎

This is due to the fact that we are calling the same method to generate
those tables than for fields label/value tables. Field label/value
tables are two columns tables with fixed width columns. Trying to put a
7 columns (matrix) or 3 columns (artifact links) in a 2 columns layout
makes MS Office or Google Docs not happy.

Closes request #39730: tables width in ttm docx generation

Change-Id: I7789974ad05e07db238e919d28a66d7dab5570f3