feat: Have an edit reference button and a form
part of story #38627 : Choice of a new Rich Text Editor

Please note that it is expected that nothing happens when you submit the
new cross reference. This contribution introduces only the UI part. The
update of an existing cross reference does not exist yet, and thus has
to be entierly implemented (big patch). To ease the review, I've
split the feature in several patches.

How to test:
- Click on an existing cross-reference
- Click on the pencil button in the popover
--> A form to edit the reference is displayed

- Update the cross reference and submit
--> The popover close
--> Nothin happens (refer to the first paragraph of the commit message)

Change-Id: I02ef9c887c329b65d17a275afcc8ae8d2502716a