feat: add automagic link
When the user writes an url and inputs a "space", it should automatically convert the url to link mark.
A plugin has been created to listen user inputs and check if it's a space and if there is a url just before the cursor, to convert it to link mark.

part of story #38627 : Choice of a new Rich Text Editor

Change-Id: I583461acccdf9d0e495b44b04f0001d71de14b1e