feat: update cross reference on submit
part of story #38627 : Choice of a new Rich Text Editor

Update strategy:

Given that we have two plugins:
- The link-popover plugin (requesting the update)
- The cross-references plugin (processing the update)

These plugins cannot directly interact, because they don't know each
other. Hence, the link-popover plugin cannot access the decorations
created by the cross-reference plugin.

The "prosemirror way" to process the update seems to be by using a
special transaction "updated-cross-reference". When this transaction is
detected by the cross-reference plugin, it will register an aditionnal
transaction whose role is to replace the old reference in the editor by
the new one (received as an EditorNode). This new reference will be
decorated by the same plugin during the apply phase happening
naturally right after the text replacement.

How to test:
- Pick up a cross reference
- Edit it and submit
--> It is updated in the editor

Change-Id: I64ea9cbb7fbc36e4b13eda723277b5a926dd1ba8