fix: request #39728: If a user does not have access at all to a tracker, it should not be displayed in the backlog
Implementation note: it seems that `testPUTCardsForReadOnlyUser` is
broken: bot user cannot access to a private project it is not member of,
it does not match the "ReadOnlyUser" in the test name. Now that we check
for tracker access for planning before checking artifact access for
card, we get a 404 instead of a 403. In order to follow latest
development about status code regarding restricted resources, we now
return a 404 in case of artifact access denied.

Change-Id: Ie99e13aa4a80c34a1e6b6b09fc8d34d1bd886141