feat: Use different program check to run top backlog post-action
Closes request #39352 Program add to top backlog post-action should
ignore Teams

How to test:
- Create a Program from the Agile Release Train XML template. Do not
link any Teams to it.
- Create a Features artifact F. Then, change its status to "To plan".
- In DB, in the table plugin_program_management_explicit_top_backlog,
you should see a row with the artifact ID of the Feature F id.
- If you deactivate the Program service in this project, the "Add to top
backlog" post-action should no longer appear in the workflow
configuration and should do nothing.
- When you add a Team to this Program, and then go to the Program app,
then you should see the Feature F in the top backlog.

Why?
The execution of the "Add to Top backlog" post-action was linked to
having a Program completely configured and linked to Teams. There is no
real functional reason to forbid this post action when the Program has
no Teams. At worst case, it will add to the top backlog, but then it
cannot be planned anywhere because there are no Teams.
This made it hard to use Templates for Programs, because by definition,
a Template of a Program will never be linked to any Teams, as its
purpose is to be duplicated.
This completes the fixes related to Template projects for Program
Management.

Notes:
PermissionBypass usage with ProgramIdentifier has been completely
removed, as this was the last usage. PermissionBypass is still used for
verifications linked to Features though.

Change-Id: I9ec7aefccba9383674d8af47e799ad3905b51552