feat: Correct reading mode weird behaviours
part of story #39685 Order artifact according to a field

How to test:
- There are no more hundreds of Vue warnings due to assignment of a
class attribute on a text/empty node. The last cells of a line still
have correct padding on the right.
- Given an existing widget with a saved report in expert mode, when you
edit the query and remove it completely, and click on "Search" (you get
an error) and then you click on "Cancel", you get back to the saved
state (with non-empty query) correctly. Before this patch, you would get
the table results but the "reading mode" with the query would stay
empty.
- When you do the same scenario but from a brand new widget, you should
correctly get the "No element selected" italic message.
- When you hover the "reading mode" with the query (and maybe the
trackers selection), the cursor should now always be "pointer" (same as
links). Previously in some cases it would keep the arrow cursor.

Change-Id: I5045a1dfecba3ed97394335d702f258d550666c7