fix: rename "embedded" to "italic"
part of story #38627 Choice of a new Rich Text Editor

The button has been wrongfully named after a confusion induced by the
`<em>` tag used to put some text in italics.

"em" does not stand for "embedded", but for "emphasized". Since it puts
the text in italics, it is better to rename the button to avoid more
confusion.

-- no functional change expected

Change-Id: Ie722f72b7d920f55f21df3a6a6579246ef3ee9c4