feat: have an edit|create image button and popover in toolbar
part of story #38627 Choice of a new Rich Text Editor

In this contribution, the image button and its popover are added to the
toolbar. It allows the user to insert a new image, or to edit an
existing one.

Note: The check contained in CheckCanInsertImage is the old check that
was made by the image plugin. It seems that it hardly prevent any
insertion, and probably needs to be reworked.

How to test:
- Click somewhere in the editor, or select a portion of text
- Click on the image button
- Enter an url and a title
- Save
--> The image has been inserted

- Click on the image you've just inserted
--> The image button is in "active" mode

- Click the image button
--> The Image source and image title fields are filled with the image
source and title

- Update the title or the source or both
- Save
--> The image has been updated

Change-Id: I66d2bd4398229dd4af921fc8f1762cd7b2f07816