fix: cache result of hash_file for avatars
Instead of computing the avatar hash for each request, we should compute
once and store it in database. There will be less load on the server and
rainforest will be grateful.

No functional change visible.

If you load a project member pages (or site admin » users), put a breakpoint in
`ComputeAvatarHash::computeAvatarHash()`, and reload the page, the
breakpoint should not be reached.

Suggested entry point for the review: `PFUser::getAvatarUrl()`

This method is now deprecated, psalm baseline is updated to whitelist
existing code that is still using it.

Introducing a DAO in piece of code that was used everywhere is not very
nice for unit tests, forcing us to inject everywhere. Sorry for the big
contribution to review 😭.

Closes request #39763 Store avatar hash

Change-Id: I3d676dc39cbcacd16282aa56a12e482a32353ff1