fix: toolbar flickering when sticky
part of story #38627 Choice of a new Rich Text Editor

What is being fixed?

In artidoc, there were a bug with the toolbar that produced flickering
and positioning issues when popover-menus where open and closed. To
avoid it, we had to move the popovers programatically outside of the
custom element, so the component height was not recomputed (it somehow
was). However, I realized that the issue was in fact due to the lack of
an explicit display rule on the webcomponent.

-- No functional change expected

Change-Id: Icb2ba6cbc6dc9f49bbdd81c23ab04e5a00b7ff89