feat: Conventional commits are handled by artifact closing references
Implements request #40063 Artifact closing on reference with conventionnal commit

The regex matching references matches context word of references with `\w+`
which not match `:`. By adding an optional `:?` the regex can now match
context words of conventional commits.

Currently handled:

- feat: -> considered as implements
- fix: -> considered as fixes
- chore: -> considered as closes

*Testing:*

Commit to a repository with this message:

    feat: art #123

It should closes artifact 123.

Change-Id: I283fc8a3de79a0f0601ab3bc6548db35c7519c98