fix: TypeError: Cannot access offset of type string on string
When a text field has the following changelog entry:

```php
[
    'fieldId'    => 'customfield_10081',
    'from'       => null,
    'fromString' => null,
    'to'         => '',
    'toString'   => 'lorem ipsum',
]
```

Then its corresponding snapshot will be mapped to a list value. This
value (an array) will be then cast as a string (since it is originaly a
text field) and PHP emits a TypeError.

Part of request #40066: TypeError: Cannot access offset of type string on string

Change-Id: I2be7ec788d92be407cc384e27d175cf3b1415bc7