fix: request #40071 Pull request e2e test fails to create a new label
No functional change expected.
make tests-e2e should pass for pullrequest e2e tests.

Why?
Escaping from cy.within() is not needed with Lazybox, because contrary
to list-picker and select2, it does not move its dropdown at the end of
<body>. The dropdown stays in the lazybox element. We can still try to
"scope" the get() calls by prefixing the lazybox selector or running a
get() before a find().
The data-test=lazybox attribute is moved to LazyboxElement itself,
previously it was on the selection element, but there are no event
listeners there.

Change-Id: Ie5e048d41e2f47d03a6816973b2d586933328b10