fix: the cursor disappears when a toolbar button is clicked
Part of story #38627 Choice of a new Rich Text Editor

What is being fixed?
When you write something in the editor and then click on a toolbar button, the editor loses focus. You have to click again in the editor, on the line you want, to write again.
Now you can write in the editor, then click on a toolbar button and continue writing without losing focus.

How to test:
-> Write in the editor, then click on 'bold' button in the toolbar and write again.
-- The text should appear bold.

Change-Id: I5f1e9fb89a91d7fb38e53e410abb29920f1c5497