fix: toolbar should not appear above the TOC
When the toolbar is stuck, there is a dropshadow that is displayed as a
clue to indicate that there is something under it.

But this dropshadow should not span above the table of content.

In order to fix this, we have to reorder zindex. This forces us to
reverts commit 3e5e4902040e600a358992860613ea16709dd948. So that toolbar
image popover is displayed above the table of content.

Same strategy is used for section dropdown, in order to display the
dropdown above the toc, its menu element is moved in the document body.

Dropdown or popover positioning zindex should not depends on z-index
nightmare of the inner application. They already have their own z-index
that is defined application wize. Maybe we should enhance createDropdown
to always move menu elements inside the document body?

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: I6caef493d558fd465145f5c3b3e0968b01aed279