fix: use standard select behavior for heading select
Usage of heading select does not match the one the user is used to.

If I take a Status selectbox, if the current value is Open, then Open is
selected. If I select Closed, then Closed is selected. Both are never disabled.

We should apply the same principle here:
* If I highlight a h1 title in the description, then h1 is selected (and not disabled)
* If I change to h2, then the highlighted section is now an h2
* If I do not change the selection and click on the same value than
  before, then nothing is changed in the editor.

This is something that has been spotted after the review of contribution
gerrit #32112 has been completed.

Part of story #38627: Choice of a new Rich Text Editor

Change-Id: If57899ea3667e917873e8222f16ae10af97463c1