fix: request #40083 Cache avatar files at the reverse proxy level
Avatar are recurrent elements in the Tuleap UI so caching them at nginx level
can save quite a lot of work to the PHP FPM worker.

Note: for instances that do not allow anonymous users to browse, seeing
the avatars require to be logged in. With this cache if you know the
full URL of the avatar you want to see you might get access to it even
if you are not logged in if someone else accessed it already. This is
an expected trade-off. Knowing the full URL means you know the SHA-256
of the avatar content, it cannot be guessed unless you already know the
avatar.

To test, redeploy nginx configuration. After the deployment, after the
first display loading the same avatar multiple times should not trigger
a call to the PHP code (the easiest way to see that is to add a log
trace somewhere early in the process). You might want to disable your
browser cache since the avatars are also cached at the client level.

Change-Id: I567215cfe48488a41b248b7b82ef628b719a4a8e