refactor: Naming in TrackerDatabaseBuilder
part of request #40099 Cross-Tracker search finishing touches 16.1

No functional change. DB integration tests should pass.

Why?
It was not obvious in "setReadPermission" that it was a method for
Tracker fields. Also the "set" implies that you can set it to true or
false. By using "grant", I'm hoping to make it clear that it gives the
permission, you cannot set it false (it means deleting from the table).
I also took the opportunity to replace "addStatusValueForArtifact" by
"buildListValue" which does the same thing.
I also used list instead of [] for types, it is stricter.

Change-Id: Id21d4eb4cafbdf116198055f54229d65224e8ee8